Add CommentSyntaxOverrides option in Scaladoc
Comment Syntax Overrides allows to override the default comment syntax
used with a chosen one based on a passed path prefix. This will allow to
use a correct syntax for specific files in one project, for example a
wiki syntax for scalalib and md syntax for dottylib.
Option argument syntax itself is similar to the one of an existing
SnippetCompiler option.